Moved symbol table to its own class
